Modified Room Entity
Room Entity now can stores days of the week into 7 bits, each represent a day of the week from Monday (MSB) to Sunday (LSB).
Hour, minute and daysofweek stored in the primitive type "short" for more efficiency because the range of "Int" is not required and overkill.
Helper functions to convert daysOfWeek from Byte to a set of DayOfWeek (java.time) and the other way too.
Using DayOfWeek for UI in RoutineVM because it is easier to use for the UI.